Coversation_bot/requirements.txt
google.genrativeai  
streamlit 
python-dotenv   